migrations, seeders and config esling with airbnb stylecode with prettier as pluggin

+++ src/db/migrations/20200617232145-create-products-table.js
@@ -17,7 +17,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.FLOAT(10, 2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       createdAt: {
